[refactor] 그림일기 페이지 연결
