+++ screens/WHTest.jsx
@@ -1,7 +1,7 @@
 import { StyleSheet, Text, View } from 'react-native'
 import React from 'react'
 import DiaryDetail from './picturediary/DiaryDetail'
-import ConsumptionSelect from './ConsumptionSelect'
+import ConsumptionSelect from './picturediary/ConsumptionSelect'
 
 const WHTest = ({navigation}) => {
   return (
